### 2022-02-10 #4146 @whedon
[PRE REVIEW]: hessQuik: Fast Hessian computation of composite functions
**Submitting author:** @elizabethnewman (<a href="http://orcid.org/0000-0002-6309-7706">Elizabeth Newman</a>)
**Repository:** <a href="https://github.com/elizabethnewman/hessQuik" target ="_blank">https://github.com/elizabethnewman/hessQuik</a>
**Version:** v0.0.1
