### 2022-02-10 #4146 @kthyng
Hi @jbytecode! Is this a reasonable fit for you?
