### 2022-02-16 #4146 @jbytecode
👋👋👋 Dear @yhtang  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**hessQuik: Fast Hessian computation of composite functions**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4146).
