### 2022-02-16 #4171 @whedon
[REVIEW]: hessQuik: Fast Hessian computation of composite functions
**Submitting author:** @elizabethnewman (<a href="http://orcid.org/0000-0002-6309-7706">Elizabeth Newman</a>)
**Repository:** https://github.com/elizabethnewman/hessQuik
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0

Related issues: https://github.com/elizabethnewman/hessQuik/issues/2
