### 2022-02-16 #4171 @jbytecode
Dear reviewers @GregaVrbancic and @yhtang

Here is the review thread. Each reviewer has 20 checkboxes with a corresponding reviewing task. Whenever you complete a task, you can manually check it. The reviewing process is continuous in JOSS so you can interact with the author and the editor during the review.

Please do not hesitate to ask anything. I will always be following the ongoing process.

Thank you in advance.

--------------------

Dear @elizabethnewman

The editorial bot suggests some DOI's for the missing bibtex items. You can start with correcting them as it suggested. You can use the test environment for compiling and examining the pdf using https://whedon.theoj.org/

Thank you in advance.
