+++ paper/paper.bib
@@ -17,6 +17,7 @@
       	eprint={2103.05180},
       	archivePrefix={arXiv},
       	primaryClass={cs.LG}
+	doi = {10.1002/gamm.202100008}
 }
 
 
@@ -162,7 +163,7 @@
         pages = {455--500}, 
         month = {September}, 
         year = {2009},
-        doi = {10.1137/07070111X},
+        doi = {10.1137/07070111X}
 }
 
 
@@ -201,11 +202,11 @@
 }
 
 @book{Goodfellow-et-al-2016,
-    title={Deep Learning},
-    author={Ian Goodfellow and Yoshua Bengio and Aaron Courville},
-    publisher={MIT Press},
-    note={\url{http://www.deeplearningbook.org}},
-    year={2016}
+    	title={Deep Learning},
+    	author={Ian Goodfellow and Yoshua Bengio and Aaron Courville},
+    	publisher={MIT Press},
+    	note={\url{http://www.deeplearningbook.org}},
+	    year={2016}
 }
 
 
