+++ paper/paper.bib
@@ -16,8 +16,8 @@
       	year={2021},
       	eprint={2103.05180},
       	archivePrefix={arXiv},
-      	primaryClass={cs.LG}
-	doi = {10.1002/gamm.202100008}
+      	primaryClass={cs.LG},
+	doi={10.1002/gamm.202100008}
 }
 
 
@@ -46,7 +46,8 @@
 	Booktitle = {Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition},
 	Pages = {770--778},
 	Title = {Deep residual learning for image recognition},
-	Year = {2016}
+	Year = {2016},
+	doi = {10.1109/cvpr.2016.90}
 }
 
 @article{Raissi:2019hv,
