+++ paper/paper.bib
@@ -57,7 +57,8 @@
 	Pages = {686--707},
 	Title = {{Physics-informed neural networks: A deep learning framework for solving forward and inverse problems involving nonlinear partial differential equations}},
 	Volume = {378},
-	Year = {2019}
+	Year = {2019},
+	doi = {10.1016/j.jcp.2018.10.045}
 }
 
 @inproceedings{huang2021convex,
