### 2022-02-16 #4171 @elizabethnewman
Dear @jbytecode

Thank you for suggesting that I correct the DOI's now.  I now have added the DOI's the editorial bot suggested.

Thank you!
