### 2022-02-16 #4171 @yhtang
I am not quite sure if it is meaningful to have two co-first authors for this two-author paper, and if this accurately reflects the contributions of the authors:
```
elizabethnewman:
225 commits
11,384 ++    6,644 --

lruthotto:
2 commits
7 ++    1 --
```
@elizabethnewman: could you please share some thoughts on it?
