### 2022-02-16 elizabethnewman/hessQuik#2 @yhtang
JOSS review: documentation
Hi, I'll use this as a master thread for issues related to the **documentation** section of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4171) of your submission "hessQuik: Fast Hessian computation of composite functions".

Below are the specific check items that are not present in the current README file of this repo:

- [x]  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
