### 2022-02-17 elizabethnewman/hessQuik#2 @elizabethnewman
Dear @yhtang

Thank you for making this master thread for documentation issues.  I have some questions regarding the four items above:

1. Is the statement of need required in the README file?  My impression was this was required for the documentation in the paper, but not in the README.  I have not seen a Statement of Need in the README file in all published JOSS papers, but perhaps I misunderstood.

2. What is missing in the README file about the functionality documentation?  I had thought the Getting Started section served this purpose, but is there something else you wanted to see?

3. We included automated tests in the tests directory https://github.com/elizabethnewman/hessQuik/tree/main/hessQuik/tests.  This is linked in relevant places in the repository.  Is this sufficient?

4. There is a Contributing section and Reporting Bugs section of the README file.  Is there something else I need to include or something unclear?

Thank you for reviewing this paper and repository!
