### 2022-02-18 #4171 @yhtang
> Dear @yhtang
>
> Thank you for your question about the contributions to the paper. The commit history does not reflect the contributions accurately. The repository was a true collaboration between me and my co-author. We had many discussions regarding the repository design, the examples to provide, and the paper writeup. It is an accurate representation of the collaboration to list us as co-first authors.

Got it. Thanks for the info!
