### 2022-02-24 elizabethnewman/hessQuik#2 @yhtang
Hi Elizabeth,

> 1. Is the statement of need required in the README file?  My impression was this was required for the documentation in the paper, but not in the README.  I have not seen a Statement of Need in the README file in all published JOSS papers, but perhaps I misunderstood.

According to the reviewer's checklist that I was given, the statement of need is required for both the *documentation* and the *paper*. Usually, the documentation of the software is hosted as a standalone site/page on places such as Github Pages or ReadTheDocs. In this case, it seems to me that the README file serves the role of the documentation for the hessQuik and hence why my initial comments targeted the README file. Definitely let me know if a documentation site is actually hosted somewhere that I neglected.

> 2. What is missing in the README file about the functionality documentation?  I had thought the Getting Started section served this purpose, but is there something else you wanted to see?

I think a common practice is to document functionality using a two-component strategy that consists of presenting a 'user guide' and an 'API documentation'. The user guide is a manual-style document that explains in natural language how the package is structured and how typical usage should occur. The API documentation is usually automatically generated using tools such as Sphinx or MkDocs to enumerate all publicly accessible methods, classes, and modules etc. Our software [FunFact](https://funfact.readthedocs.io/en/latest/) might serve as a (hopefully not too terrible) example of this pattern.

> 3. We included automated tests in the tests directory https://github.com/elizabethnewman/hessQuik/tree/main/hessQuik/tests.  This is linked in relevant places in the repository.  Is this sufficient?

I did find the tests. Maybe this is my misunderstanding, but I was additionally looking for a mechanism that automatically executes the tests every time when new changes are introduced to the code. This is usually implemented using continuous integration services such as Github Actions or Travis CI or similar things. I suggest we fall back to the main review thread and consult the editor on this matter.

> 4. There is a Contributing section and Reporting Bugs section of the README file.  Is there something else I need to include or something unclear?

This is probably my bad that I missed this section initially. Let's check this box now :grin:
