### 2022-02-24 #4171 @yhtang
I think we need some clarification regarding what suffices as 'automated tests' following up on the discussion [here](https://github.com/elizabethnewman/hessQuik/issues/2#issuecomment-1049563899). Specifically, are continuous integration setups required in addition to having unit tests written?
