### 2022-02-24 #4171 @jbytecode
Formally we always check the software if it has unit tests with an acceptable level of coverage in JOSS.

Despite it is not a rigid requirement, I suggest adding GitHub Actions and some other continuous integration tools. In our case, yes, it would be better to have CI on GitHub so we can track if the latest changes don't break the code.
