### 2022-02-24 #4171 @danielskatz
See https://joss.readthedocs.io/en/latest/review_criteria.html#tests for more guidance
