### 2022-02-25 a07241b
Update README.md
added statement of need to README
