### 2022-02-25 0bdbc41
Update hessQuik_activation_function.py
added some functionality documentation
