+++ hessQuik/activations/hessQuik_activation_function.py
@@ -11,6 +11,27 @@ class hessQuikActivationFunction(nn.Module):
 
     def forward(self, x: Tensor, do_gradient: bool = False, do_Hessian: bool = False, forward_mode: bool = True) -> \
             Tuple[Tensor, Union[Tensor, None], Union[Tensor, None]]:
+                """
+        Forward propagate through a pointwise actication function, sigma(x)
+        
+        Parameters
+        ----------
+        x : (N, in_features) torch.Tensor
+            Tensor of input features
+        do_gradient : bool, optional
+            If True, compute the gradient of the output of the layer, sigma(x), with respect to the layer input, x
+        do_Hessian : bool, optional
+            If True, compute the Hessian of the output of the layer, sigma(x), with respect to the layer input, x
+            
+        Returns
+        -------
+        sigma : (N, in_features) torch.Tensor
+            Output of the layer, sigma(x)
+        dsigma : (N, in_features)
+            Gradient of the output of the layer, sigma(x), with respect to the layer inputs, x
+        d2sigma : (N, in_features)
+            Hessian of the output of the layer, sigma(x), with respect to the layer inputs, x
+        """
         raise NotImplementedError
 
     def backward(self, do_Hessian: bool = False) -> Tuple[Tensor, Union[Tensor, None]]:
