### 2022-02-25 elizabethnewman/hessQuik#2 @GregaVrbancic
Hi Elizabeth,

I am the other reviewer of your submission. Overall, I agree with @yhtang comments regarding the documentation. In current state could be a bit confusing since and hard to find, since there are bits of documentation all over the project structure. I would appreciate at least some navigation in the main README file, so that one can navigate to the specific part of documentation more easily.

Regarding the tests, I have no problem with manual tests as long there is a brief description on how to run them. That is also in line with [JOSS review criteria.](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)
