### 2022-02-25 elizabethnewman/hessQuik#2 @GregaVrbancic
Looking more in depth I see the major problem, regarding the documentation, that API method documentation is missing. In my opinion, this concern should be addressed first.
