### 2022-02-25 elizabethnewman/hessQuik#2 @elizabethnewman
Dear @yhtang and @GregaVrbancic,

Thank you for your feedback about the repository and for answering my questions - it is very helpful.  I will work on clarifying the API method documentation and see if I can add automatic tests.  The latter does not seem as straightforward due to the nature of this repository (tests often need to be tailored to the specific layer or network), but I will certainly do my best!
