+++ hessQuik/activations/hessQuik_activation_function.py
@@ -11,9 +11,9 @@ class hessQuikActivationFunction(nn.Module):
 
     def forward(self, x: Tensor, do_gradient: bool = False, do_Hessian: bool = False, forward_mode: bool = True) -> \
             Tuple[Tensor, Union[Tensor, None], Union[Tensor, None]]:
-                """
-        Forward propagate through a pointwise actication function, sigma(x)
-        
+        """
+        Forward propagate through a pointwise activation function, sigma(x)
+
         Parameters
         ----------
         x : (N, in_features) torch.Tensor
@@ -22,7 +22,7 @@ class hessQuikActivationFunction(nn.Module):
             If True, compute the gradient of the output of the layer, sigma(x), with respect to the layer input, x
         do_Hessian : bool, optional
             If True, compute the Hessian of the output of the layer, sigma(x), with respect to the layer input, x
-            
+
         Returns
         -------
         sigma : (N, in_features) torch.Tensor
@@ -32,12 +32,19 @@ class hessQuikActivationFunction(nn.Module):
         d2sigma : (N, in_features)
             Hessian of the output of the layer, sigma(x), with respect to the layer inputs, x
         """
+
         raise NotImplementedError
 
     def backward(self, do_Hessian: bool = False) -> Tuple[Tensor, Union[Tensor, None]]:
+        """
+        This is the backward method
+        """
         dsigma, d2sigma = self.compute_derivatives(*self.ctx, do_Hessian=do_Hessian)
 
         return dsigma, d2sigma
 
     def compute_derivatives(self, *args, do_Hessian: bool = False) -> Tuple[Tensor, Union[Tensor, None]]:
+        """
+        This is the compute derivatives method
+        """
         raise NotImplementedError
