+++ docs/build/html/_sources/hessQuik_API/hessQuik.activations.rst.txt
@@ -1,0 +1,65 @@
+hessQuik Activation Functions
+=============================
+
+hessQuik.activations.hessQuik\_activation\_function
+---------------------------------------------------
+
+.. automodule:: hessQuik.activations.hessQuik_activation_function
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.activations.antiTanh\_activation
+-----------------------------------------
+
+.. autoclass:: hessQuik.activations.antiTanh_activation.antiTanhActivation
+   :members:
+   :no-undoc-members:
+
+hessQuik.activations.identity\_activation
+-----------------------------------------
+
+.. automodule:: hessQuik.activations.identity_activation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.activations.quadratic\_activation
+------------------------------------------
+
+.. automodule:: hessQuik.activations.quadratic_activation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.activations.sigmoid\_activation
+----------------------------------------
+
+.. automodule:: hessQuik.activations.sigmoid_activation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.activations.softplus\_activation
+-----------------------------------------
+
+.. automodule:: hessQuik.activations.softplus_activation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.activations.tanh\_activation
+-------------------------------------
+
+.. automodule:: hessQuik.activations.tanh_activation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: hessQuik.activations
+   :members:
+   :undoc-members:
+   :show-inheritance:
