+++ docs/build/html/_sources/hessQuik_API/hessQuik.layers.rst.txt
@@ -1,0 +1,61 @@
+hessQuik.layers package
+=======================
+
+Submodules
+----------
+
+hessQuik.layers.hessQuik\_layer module
+--------------------------------------
+
+.. automodule:: hessQuik.layers.hessQuik_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.layers.icnn\_layer module
+----------------------------------
+
+.. automodule:: hessQuik.layers.icnn_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.layers.quadratic\_icnn\_layer module
+---------------------------------------------
+
+.. automodule:: hessQuik.layers.quadratic_icnn_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.layers.quadratic\_layer module
+---------------------------------------
+
+.. automodule:: hessQuik.layers.quadratic_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.layers.resnet\_layer module
+------------------------------------
+
+.. automodule:: hessQuik.layers.resnet_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.layers.single\_layer module
+------------------------------------
+
+.. automodule:: hessQuik.layers.single_layer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: hessQuik.layers
+   :members:
+   :undoc-members:
+   :show-inheritance:
