+++ docs/build/html/_sources/getting_started.rst.txt
@@ -1,0 +1,4 @@
+Getting Started
+===============
+
+Here is how you use the code.
