+++ docs/build/html/_sources/hessQuik_API/hessQuik.rst.txt
@@ -1,0 +1,21 @@
+hessQuik package
+================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   hessQuik.activations
+   hessQuik.layers
+   hessQuik.networks
+   hessQuik.utils
+
+Module contents
+---------------
+
+.. automodule:: hessQuik
+   :members:
+   :undoc-members:
+   :show-inheritance:
