+++ docs/build/html/_sources/hessQuik_API/hessQuik.utils.rst.txt
@@ -1,46 +1,0 @@
-hessQuik.utils package
-======================
-
-Functions
----------
-
-hessQuik.utils.data
---------------------
-
-.. py:function:: hessQuik.utils.data
-
-hessQuik.utils.input\_derivative\_check
----------------------------------------
-
-.. py:function:: hessQuik.utils.input_derivative_check
-
-hessQuik.utils.network\_derivative\_check
------------------------------------------
-
-.. py:function:: hessQuik.utils.network_derivative_check
-
-hessQuik.utils.timing
----------------------
-
-.. py:function:: hessQuik.utils.timing
-
-hessQuik.utils.training
------------------------
-
-.. py:function:: hessQuik.utils.training
-
-hessQuik.utils.utils
---------------------
-
-.. automodule:: hessQuik.utils.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: hessQuik.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
