+++ docs/build/html/_sources/hessQuik_API/index.rst.txt
@@ -1,13 +1,0 @@
-hessQuik API
-============
-
-.. toctree::
-   :maxdepth: 2
-   :caption: hessQuik API
-
-   hessQuik.activations
-   hessQuik.layers
-   hessQuik.networks
-   hessQuik
-   hessQuik.utils
-   modules
