+++ hessQuik/__init__.py
@@ -1,1 +1,0 @@
-# TODO: should we import torch here?
