+++ hessQuik/activations2/__init__.py
@@ -1,0 +1,1 @@
+from .hessQuik_activation_function2 import hessQuikActivationFunction
