+++ docs/source/hessQuik_API/hessQuik.activations2.rst
@@ -1,0 +1,10 @@
+hessQuik Activation Functions 2
+===============================
+
+hessQuik.activations2.hessQuik\_activation\_function
+----------------------------------------------------
+
+.. automodule:: hessQuik.activations2.hessQuik_activation_function2
+   :members:
+   :undoc-members:
+   :show-inheritance:
