+++ docs/source/getting_started.rst
@@ -1,4 +1,0 @@
-Getting Started
-===============
-
-Here is how you use the code.
