+++ docs/source/index.rst
@@ -1,28 +1,0 @@
-.. hessQuik documentation master file, created by
-   sphinx-quickstart on Fri Feb 25 21:20:58 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to hessQuik's documentation!
-====================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Overview
-
-   getting_started
-
-
-.. toctree::
-   :caption: hessQuik API
-   :glob:
-
-   hessQuik_API/index
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
