+++ docs/source/hessQuik_API/modules.rst
@@ -1,7 +1,0 @@
-hessQuik
-========
-
-.. toctree::
-   :maxdepth: 4
-
-   hessQuik
