+++ docs/source/hessQuik_API/index.rst
@@ -1,14 +1,0 @@
-hessQuik API
-============
-
-.. toctree::
-   :maxdepth: 2
-   :caption: hessQuik API
-
-   hessQuik.activations
-   hessQuik.activations2
-   hessQuik.layers
-   hessQuik.networks
-   hessQuik
-   hessQuik.utils
-   modules

[…]
