### 2022-02-26 e1360cc
new documentation for antiTanh
