+++ hessQuik/activations/antiTanh_activation.py
@@ -3,13 +3,21 @@ from hessQuik.activations import hessQuikActivationFunction
 
 
 class antiTanhActivation(hessQuikActivationFunction):
-    """
-    Here is some documentation.
-    """
+
     def __init__(self):
         super(antiTanhActivation, self).__init__()
 
     def forward(self, x, do_gradient=False, do_Hessian=False, forward_mode=True):
+        """
+        .. math::
+
+            \begin{align}
+                \sigma(x)   & = \ln(\cosh(x))\\
+                \sigma'(x)  & = \tanh(x)\\
+                \sigma''(x) & = 1 - \tanh^2(x)
+            \end{align}
+        """
+        
         (dsigma, d2sigma) = (None, None)
 
         # forward
