+++ hessQuik/activations/antiTanh_activation.py
@@ -6,7 +6,7 @@ class antiTanhActivation(hessQuikActivationFunction):
     """
     antiTanh Activation function
 
-    :math::
+    .. math::
         \sigma(x)   & = \ln(\cosh(x))\\
         \sigma'(x)  & = \tanh(x)\\
         \sigma''(x) & = 1 - \tanh^2(x)
