+++ hessQuik/activations/antiTanh_activation.py
@@ -3,17 +3,14 @@ from hessQuik.activations import hessQuikActivationFunction
 
 
 class antiTanhActivation(hessQuikActivationFunction):
-    """
+    r"""
     antiTanh Activation function
 
     .. math::
-
-        \sigma(x) = \ln(\cosh(x))
-
-        \sigma'(x) = \tanh(x)
-
-        \sigma''(x) = 1 - \tanh^2(x)
-
+        :nowrap:
+            \begin{align}
+                \sigma(x) &= \ln(\cosh(x))
+            \end{align}
 
     """
 
