### 2022-02-26 273b9cb
new activations documentation
