+++ hessQuik/activations/sigmoid_activation.py
@@ -3,11 +3,26 @@ from hessQuik.activations import hessQuikActivationFunction
 
 
 class sigmoidActivation(hessQuikActivationFunction):
+    r"""
+    Sigmoid function
+
+    .. math::
+
+        \begin{align}
+            \sigma(x)   &= \frac{1}{1 + e^{-x}}\\
+            \sigma'(x)  &= \sigma(x)(1 - \sigma(x))\\
+            \sigma''(x) &= \sigma'(x)(1 - 2 * sigma)
+        \end{align}
+
+    """
 
     def __init__(self):
         super(sigmoidActivation, self).__init__()
 
     def forward(self, x, do_gradient=False, do_Hessian=False, forward_mode=True):
+        """
+        :meta private:
+        """
         (dsigma, d2sigma) = (None, None)
 
         # forward propagate
@@ -23,6 +38,9 @@ class sigmoidActivation(hessQuikActivationFunction):
         return sigma, dsigma, d2sigma
 
     def compute_derivatives(self, *args, do_Hessian=False):
+        """
+        :meta private:
+        """
         sigma = args[0]
         dsigma = sigma * (1 - sigma)
         d2sigma = None
