+++ hessQuik/activations/tanh_activation.py
@@ -4,17 +4,10 @@ from hessQuik.activations import hessQuikActivationFunction
 
 class tanhActivation(hessQuikActivationFunction):
     r"""
-    Hyperbolic tangent function
-
-    .. math::
-
-        \begin{align}
-            \sigma(x)   &= \tanh(x)\\
-            \sigma'(x)  &= 1 - \tanh^2(x)\\
-            \sigma''(x) &= -2\tanh(x) (1 - \tanh^2(x))
-        \end{align}
+    Applies the hyperbolic tangent activation function to each entry of the incoming data.
 
     Examples::
+
         >>> act_func = act.tanhActivation()
         >>> x = torch.randn(10, 4)
         >>> sigma, dsigma, d2sigma = act_func(x)
@@ -26,8 +19,12 @@ class tanhActivation(hessQuikActivationFunction):
         super(tanhActivation, self).__init__()
 
     def forward(self, x, do_gradient=False, do_Hessian=False, forward_mode=True):
-        """
-        :meta private:
+        r"""
+        Applies the hyperbolic tangent activation function to each entry of the incoming data.
+
+        .. math::
+
+            \sigma(x)  = \tanh(x)
         """
         (dsigma, d2sigma) = (None, None)
 
@@ -44,8 +41,15 @@ class tanhActivation(hessQuikActivationFunction):
         return sigma, dsigma, d2sigma
 
     def compute_derivatives(self, *args, do_Hessian=False):
-        """
-        :meta private:
+        r"""
+        Computes the first and second derivatives of each entry of the incoming data.
+
+        .. math::
+            \begin{align}
+                \sigma'(x)  &= 1 - \tanh^2(x)\\
+                \sigma''(x) &= -2\tanh(x) (1 - \tanh^2(x))
+            \end{align}
+
         """
         sigma = args[0]
         d2sigma = None
