+++ hessQuik/activations/hessQuik_activation_function.py
@@ -18,28 +18,35 @@ class hessQuikActivationFunction(nn.Module):
         Applies a pointwise activation function to the incoming data.
 
         Input:
+
             x (torch.Tensor): input into the activation function. :math:`(*)` where :math:`*` means any size.
+
             do_gradient (bool): If set to ``True``, the gradient will be computed during the forward call.
                                 Default: ``True``
+
             do_Hessian (bool): If set to ``True``, the Hessian will be computed during the forward call.
                                 Default: ``True``
+
             forward_mode (bool): If set to ``False``, the derivatives will be computed in backward mode.
                                 Default: ``True``
 
         Return:
+
             sigma (torch.Tensor): value of activation function at input x, same size as x
+
             dsigma (torch.Tensor): optional, first derivative of activation function at input x, same size as x
+
             d2sigma (torch.Tensor): optional, second derivative of activation function at input x, same size as x
         """
         raise NotImplementedError
 
     def backward(self, do_Hessian: bool = False) -> Tuple[Tensor, Union[Tensor, None]]:
         r"""
-        Computes derivatives of activation function evaluated at x in backward mode
+        Computes derivatives of activation function evaluated at x in backward mode.
 
-        Calls self.compute_derivatives without inputs, stores necessary variables in self.ctx
+        Calls self.compute_derivatives without inputs, stores necessary variables in self.ctx.
 
-        Inherited by all subclasses
+        Inherited by all subclasses.
         """
         dsigma, d2sigma = self.compute_derivatives(*self.ctx, do_Hessian=do_Hessian)
 
@@ -50,12 +57,16 @@ class hessQuikActivationFunction(nn.Module):
         Computes derivatives of activation function evaluated at x in either forward or backward more
 
         Input:
+
             *args (tuple): tuple of variables needed to compute derivatives
+
             do_Hessian (bool): If set to ``True``, the Hessian will be computed during the forward call.
                                 Default: ``True``
 
         Return:
+            
             dsigma (torch.Tensor): optional, first derivative of activation function at input x, same size as x
+
             d2sigma (torch.Tensor): optional, second derivative of activation function at input x, same size as x
         """
         raise NotImplementedError
