+++ hessQuik/activations/hessQuik_activation_function.py
@@ -16,7 +16,7 @@ class hessQuikActivationFunction(nn.Module):
             -> Tuple[Tensor, Union[Tensor, None], Union[Tensor, None]]:
         r"""
         Applies a pointwise activation function to the incoming data.
-        
+
         :param x: input into the activation function. :math:`(*)` where :math:`*` means any size.
         :type x: torch.Tensor
         :param do_gradient: If set to ``True``, the gradient will be computed during the forward call. Default: ``False``
@@ -25,8 +25,10 @@ class hessQuikActivationFunction(nn.Module):
         :type do_Hessian: bool, optional
         :param forward_mode:  If set to ``False``, the derivatives will be computed in backward mode. Default: ``True``
         :type forward_mode: bool, optional
-        :return: sigma: value of activation function at input x, same size as x
-        :rtype: torch.Tensor
+        :return:
+            -sigma (:py:class:`torch.Tensor`) - value of activation function at input x, same size as x
+            -dsigma (:py:class:`Union[torch.Tensor, None]`, optional) - first derivative of activation function at input x, same size as x
+            -d2sigma (:py:class:`torch.Tensor`, optional) - second derivative of activation function at input x, same size as x
 
         """
         raise NotImplementedError
