### 2022-02-27 ea47eac
updated activation documentation
