+++ hessQuik/activations/identity_activation.py
@@ -4,15 +4,14 @@ from hessQuik.activations import hessQuikActivationFunction
 
 class identityActivation(hessQuikActivationFunction):
     r"""
-    Identity function
+    Applies the identity activation function to each entry of the incoming data.
 
-    .. math::
+    Examples::
 
-        \begin{align}
-            \sigma(x)   &= x\\
-            \sigma'(x)  &= 1\\
-            \sigma''(x) &= 0
-        \end{align}
+        >>> import hessQuik.activations as act
+        >>> act_func = act.identityActivation()
+        >>> x = torch.randn(10, 4)
+        >>> sigma, dsigma, d2sigma = act_func(x, do_gradient=True, do_Hessian=True)
 
     """
 
@@ -20,8 +19,12 @@ class identityActivation(hessQuikActivationFunction):
         super(identityActivation, self).__init__()
 
     def forward(self, x, do_gradient=False, do_Hessian=False, forward_mode=True):
-        """
-        :meta private:
+        r"""
+        Activates each entry of incoming data via
+
+        .. math::
+
+            \sigma(x)  = x
         """
         (dsigma, d2sigma) = (None, None)
 
@@ -38,8 +41,15 @@ class identityActivation(hessQuikActivationFunction):
         return sigma, dsigma, d2sigma
 
     def compute_derivatives(self, *args, do_Hessian=False):
-        """
-        :meta private:
+        r"""
+        Computes the first and second derivatives of each entry of the incoming data via
+
+        .. math::
+            \begin{align}
+                \sigma'(x)  &= 1\\
+                \sigma''(x) &= 0
+            \end{align}
+
         """
         x = args[0]
         d2sigma = None
