### 2022-02-27 104062f
new hessQuik_layer documentation
