+++ hessQuik/layers/hessQuik_layer.py
@@ -31,19 +31,14 @@ class hessQuikLayer(nn.Module):
                 dudx: Union[Tensor, None] = None, d2ud2x: Union[Tensor, None] = None) \
             -> Tuple[Tensor, Union[Tensor, None], Union[Tensor, None]]:
         r"""
-        Forward pass through the layer that maps
-        .. math::
-
-            \text{input features :math:`u` of size :math:`(n_s, n_{in})`} \longleftarrow
-            output features :math:`f` of size :math:`(n_s, n_{out})`
-
-        where :math:`n_s` is the number of samples and :math:`n_{in}` is the number of input features,
-        and :math:`n_{out}` is the number of output features.
+        Forward pass through the layer that maps input features :math:`u` of size :math:`(n_s, n_{in})`
+        to output features :math:`f` of size :math:`(n_s, n_{out})` where :math:`n_s` is the number of samples,
+         :math:`n_{in}` is the number of input features, and :math:`n_{out}` is the number of output features.
 
         The input features :math:`u(x)` is a function of the network input :math:`x` of size :math:`(n_s, d)`
         where :math:`d` is the dimension of the network input.
 
-        :param u: features from previous layer. :math:`(n_s, n_{in})`
+        :param u: features from previous layer with shape :math:`(n_s, n_{in})`
         :type u: torch.Tensor
         :param do_gradient: If set to ``True``, the gradient will be computed during the forward call. Default: ``False``
         :type do_gradient: bool, optional
@@ -51,15 +46,15 @@ class hessQuikLayer(nn.Module):
         :type do_Hessian: bool, optional
         :param forward_mode:  If set to ``False``, the derivatives will be computed in backward mode. Default: ``True``
         :type forward_mode: bool, optional
-        :param dudx: gradient of features from previous layer with respect to network input x. :math:`(n_s, d, n_{in})`
+        :param dudx: if ``forward_mode = True``, gradient of features from previous layer with respect to network input :math:`x` with shape :math:`(n_s, d, n_{in})`
         :type dudx: torch.Tensor or ``None``
-        :param d2ud2x: Hessian of features from previous layer with respect to network input x. :math:`(n_s, d, d, n_{in})`
+        :param d2ud2x: if ``forward_mode = True``, Hessian of features from previous layer with respect to network input :math:`x` with shape :math:`(n_s, d, d, n_{in})`
         :type d2ud2x: torch.Tensor or ``None``
         :return:
 
-            - **f** (*torch.Tensor*) - output features of layer. :math:`(n_s, n_{out})`
-            - **df** (*torch.Tensor* or ``None``) - if ``forward_mode = True``, gradient of output features with respect to network input x. :math:`(n_s, d, n_{out})`
-            - **d2f** (*torch.Tensor* or ``None``) - if ``forward_mode = True``, Hessian of output features with respect to network input x. :math:`(n_s, d, d, n_{out})`
+            - **f** (*torch.Tensor*) - output features of layer with shape :math:`(n_s, n_{out})`
+            - **dfdx** (*torch.Tensor* or ``None``) - if ``forward_mode = True``, gradient of output features with respect to network input :math:`x` with shape :math:`(n_s, d, n_{out})`
+            - **d2fd2x** (*torch.Tensor* or ``None``) - if ``forward_mode = True``, Hessian of output features with respect to network input :math:`x` with shape :math:`(n_s, d, d, n_{out})`
 
         """
 
@@ -68,25 +63,26 @@ class hessQuikLayer(nn.Module):
     def backward(self, do_Hessian: bool = False,
                  dgdf: Union[Tensor, None] = None, d2gd2f: Union[Tensor, None] = None) \
             -> Tuple[Tensor, Union[Tensor, None]]:
[…]
