+++ hessQuik/layers/__init__.py
@@ -1,4 +1,3 @@
-
 from .hessQuik_layer import hessQuikLayer
 from .single_layer import singleLayer
 from .resnet_layer import resnetLayer
