### 2022-02-27 2124437
new resnet_layer documentation
