+++ docs/hessQuik_setup/index.rst
@@ -1,0 +1,7 @@
+Getting Started
+===============
+
+.. toctree::
+   :maxdepth: 2
+
+   install
