+++ docs/hessQuik_API/index.rst
@@ -1,0 +1,10 @@
+hessQuik API
+============
+
+.. toctree::
+   :maxdepth: 2
+
+   hessQuik.activations
+   hessQuik.layers
+   hessQuik.networks
+
