+++ docs/index.rst
@@ -1,0 +1,28 @@
+.. Elizabeth Newman and Lars Ruthotto documentation master file, created by
+   sphinx-quickstart on Sat Feb 26 12:45:02 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to hessQuik's documentation!
+==============================================================
+
+.. toctree::
+   :caption: Overview
+   :glob:
+
+   hessQuik_setup/index
+
+.. toctree::
+   :caption: Documentation
+   :glob:
+
+   hessQuik_API/index
+   hessQuik_derivations/index
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
