+++ docs/hessQuik_API/hessQuik.networks.rst
@@ -1,0 +1,45 @@
+hessQuik.networks package
+=========================
+
+Submodules
+----------
+
+hessQuik.networks.fully\_connected\_network module
+--------------------------------------------------
+
+.. automodule:: hessQuik.networks.fully_connected_network
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.networks.hessQuik\_network module
+------------------------------------------
+
+.. automodule:: hessQuik.networks.hessQuik_network
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.networks.icnn\_network module
+--------------------------------------
+
+.. automodule:: hessQuik.networks.icnn_network
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+hessQuik.networks.resnet\_network module
+----------------------------------------
+
+.. automodule:: hessQuik.networks.resnet_network
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: hessQuik.networks
+   :members:
+   :undoc-members:
+   :show-inheritance:
