+++ docs/hessQuik_API/hessQuik.activations.rst
@@ -1,0 +1,77 @@
+hessQuik.activations
+=====================
+
+hessQuik Activation Function
+----------------------------
+
+.. automodule:: hessQuik.activations.hessQuik_activation_function
+   :members:
+   :show-inheritance:
+
+AntiTanh
+--------
+
+.. automodule:: hessQuik.activations.antiTanh_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/antiTanh.png
+  :width: 400
+  :align: center
+
+Identity
+--------
+
+.. automodule:: hessQuik.activations.identity_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/identity.png
+  :width: 400
+  :align: center
+
+Quadratic
+---------
+
+.. automodule:: hessQuik.activations.quadratic_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/quadratic.png
+  :width: 400
+  :align: center
+
+
+Sigmoid
+-------
+
+.. automodule:: hessQuik.activations.sigmoid_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/sigmoid.png
+  :width: 400
+  :align: center
+
+Softplus
+--------
+
+.. automodule:: hessQuik.activations.softplus_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/softplus.png
+  :width: 400
+  :align: center
+
+Tanh
+----
+
+.. automodule:: hessQuik.activations.tanh_activation
+   :members:
+   :show-inheritance:
+
+.. image:: ../img/activations/tanh.png
+  :width: 400
+  :align: center
+
