+++ docs/hessQuik_API/hessQuik.layers.rst
@@ -1,0 +1,48 @@
+hessQuik.layers
+===============
+
+hessQuik Layer
+--------------
+
+.. automodule:: hessQuik.layers.hessQuik_layer
+   :members:
+   :show-inheritance:
+
+ICNN Layer
+----------
+
+.. automodule:: hessQuik.layers.icnn_layer
+   :members:
+   :show-inheritance:
+
+Quadratic ICNN Layer
+--------------------
+
+.. automodule:: hessQuik.layers.quadratic_icnn_layer
+   :members:
+   :show-inheritance:
+
+Quadratic Layer
+---------------
+
+.. automodule:: hessQuik.layers.quadratic_layer
+   :members:
+   :show-inheritance:
+
+Residual Layer
+--------------
+
+.. automodule:: hessQuik.layers.resnet_layer
+   :members:
+   :special-members: __init__
+   :show-inheritance:
+
+
+Single Layer
+------------
+
+.. automodule:: hessQuik.layers.single_layer
+   :members:
+   :special-members: __init__
+   :show-inheritance:
+
