+++ docs/hessQuik_setup/install.rst
@@ -1,0 +1,6 @@
+Installation
+============
+
+Use ``pip`` from the command line::
+
+    python -m pip install git+https://github.com/elizabethnewman/hessQuik.git

[…]
