### 2022-02-27 13cf694
new icnn layer documentation
