+++ docs/conf.py
@@ -13,6 +13,7 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
+autodoc_mock_imports = ['torch']
 
 
 # -- Project information -----------------------------------------------------
