+++ hessQuik/layers/icnn_layer.py
@@ -95,13 +95,13 @@ class ICNNLayer(hessQuikLayer):
 
         .. math::
 
-            \left[\begin{array}{c}f(x) & x\end{array}\right] =
-            \sigma\left(\left[\begin{array}{c}u(x) & x\end{array}\right]
-            \left[\begin{array}{c}L^+ \\ K\end{array}\right] + b\right)
+            f(x) =
+            \left[\begin{array}{c} \sigma\left(\left[\begin{array}{c}u(x) & x\end{array}\right]
+            \left[\begin{array}{c}L^+ \\ K\end{array}\right] + b\right) & x \end{array}\right]
 
         Here, :math:`u(x)` is the input into the layer of size :math:`(n_s, n_{in})` which is
         a function of the input of the network, :math:`x` of size :math:`(n_s, d)`.
-        The output features, :math:`f(x)`, are of size :math:`(n_s, n_{out})`.
+        The output features, :math:`f(x)`, are of size :math:`(n_s, n_{out} + d)`.
         The notation :math:`(\cdot)^+` is a function that makes the weights of a matrix nonnegative.
 
         As an example, for one sample, :math:`n_s = 1`, the gradient with respect to :math:`x` is of the form
@@ -115,6 +115,7 @@ class ICNNLayer(hessQuikLayer):
 
         where :math:`\text{diag}` transforms a vector into the entries of a diagonal matrix and :math:`I` is
         the :math:`d \times d` identity matrix.
+
         """
 
         (dfdx, d2fd2x) = (None, None)
@@ -171,9 +172,9 @@ class ICNNLayer(hessQuikLayer):
 
         .. math::
 
-            \left[\begin{array}{c}f(u) & x\end{array}\right] =
-            \sigma\left(\left[\begin{array}{c}u & x\end{array}\right]
-            \left[\begin{array}{c}L^+ \\ K\end{array}\right] + b\right)
+            f(u) =
+            \left[\begin{array}{c} \sigma\left(\left[\begin{array}{c}u & x\end{array}\right]
+            \left[\begin{array}{c}L^+ \\ K\end{array}\right] + b\right) & x \end{array}\right]
 
         Here, the network is :math:`g` is a function of :math:`f(u)`.
 
@@ -192,49 +193,49 @@ class ICNNLayer(hessQuikLayer):
             M = torch.cat((self.nonneg(self.L), M), dim=0)
 
         # obtain stored information from backward pass
-        d2gd2x = None
+        d2gd2u = None
         dsig, d2sig = self.act.backward(do_Hessian=do_Hessian)
 
         # compute gradient
-        dgdx = dsig.unsqueeze(1) * M
+        dgdu = dsig.unsqueeze(1) * M
 
         # augment gradient
-        M2 = torch.ones(dgdx.shape[0], 1, 1, dtype=dgdx.dtype, device=dgdx.device) \
-            * torch.eye(self.input_dim, dtype=dgdx.dtype, device=dgdx.device).unsqueeze(0)
+        M2 = torch.ones(dgdu.shape[0], 1, 1, dtype=dgdu.dtype, device=dgdu.device) \
+            * torch.eye(self.input_dim, dtype=dgdu.dtype, device=dgdu.device).unsqueeze(0)
 
         if self.in_features is not None:
-            Z = torch.zeros(dgdx.shape[0], self.input_dim, self.in_features)
+            Z = torch.zeros(dgdu.shape[0], self.input_dim, self.in_features)
             M2 = torch.cat((Z, M2), dim=-1).permute(0, 2, 1)
 
[…]
