+++ docs/conf.py
@@ -62,7 +62,8 @@ html_theme_options = {'sticky_navigation': True}
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+# html_static_path = ['_static']
+html_static_path = []
 
 # order of functions
 autodoc_member_order = 'bysource'
