### 2022-02-27 b94b897
complete draft of layer documentation
