+++ docs/index.rst
@@ -7,12 +7,14 @@ Welcome to hessQuik's documentation!
 ==============================================================
 
 .. toctree::
+   :maxdepth: 2
    :caption: Overview
    :glob:
 
    hessQuik_setup/index
 
 .. toctree::
+   :maxdepth: 2
    :caption: Documentation
    :glob:
 
