+++ docs/hessQuik_derivations/index.rst
@@ -1,8 +1,9 @@
-hessQuik Derivations
-====================
+hessQuik Derivations (coming soon!)
+===================================
 
 .. toctree::
    :maxdepth: 2
+   :Caption: hessQuik Derivations (coming soon!)
 
    hessQuik.singleLayer_derivation
 
