+++ docs/hessQuik_API/index.rst
@@ -3,6 +3,7 @@ hessQuik API
 
 .. toctree::
    :maxdepth: 2
+   :Caption: hessQuik API
 
    hessQuik.activations
    hessQuik.layers
