+++ hessQuik/utils/timing.py
@@ -4,10 +4,13 @@ import hessQuik.layers as lay
 import hessQuik.networks as net
 import time
 import gc
+from typing import Union
 
 
-def setup_device_and_gradient(f, network_wrapper='hessQuik', device='cpu'):
-
+def setup_device_and_gradient(f: net.NN, network_wrapper: str = 'hessQuik', device: str = 'cpu'):
+    r"""
+    Setup network with correct wrapper and device
+    """
     # map to device
     f = f.to(device)
 
@@ -20,7 +23,10 @@ def setup_device_and_gradient(f, network_wrapper='hessQuik', device='cpu'):
     return f
 
 
-def setup_resnet(in_features, out_features, width=20, depth=4):
+def setup_resnet(in_features: int, out_features: int, width: int = 16, depth: int = 4) -> net.NN:
+    r"""
+    Setup resnet architecture for timing tests
+    """
 
     f = net.NN(lay.singleLayer(in_features, width, act=act.antiTanhActivation()),
                net.resnetNN(width, depth, h=0.5, act=act.tanhActivation()),
@@ -28,12 +34,21 @@ def setup_resnet(in_features, out_features, width=20, depth=4):
     return f
 
 
-def setup_fully_connected(in_features, out_features, width=20, depth=4):
+def setup_fully_connected(in_features: int, out_features: int, width: int = 16, depth: int = 4):
+    r"""
+    Setup fully-connected architecture for timing tests
+    """
     f = net.fullyConnectedNN([in_features] + depth * [width] + [out_features], act=act.tanhActivation())
     return f
 
 
-def setup_icnn(in_features, out_features, width=20, depth=4):
+def setup_icnn(in_features: int, out_features: int, width: int = 16, depth: int = 4):
+    r"""
+    Setup ICNN architecture for timing tests.
+
+    Requires scalar output.
+    """
+
     if out_features > 1:
         raise ValueError('Must be scalar output for ICNN example')
 
@@ -42,8 +57,11 @@ def setup_icnn(in_features, out_features, width=20, depth=4):
     return f
 
 
-def setup_network(in_features, out_features, width, depth, network_type='resnet',
-                  network_wrapper='hessQuik', device='cpu'):
+def setup_network(in_features: int, out_features: int, width: int, depth: int, network_type: str = 'resnet',
+                  network_wrapper: str = 'hessQuik', device: str = 'cpu'):
+    r"""
+    Wrapper to setup network.
+    """
     if network_type == 'resnet':
         f = setup_resnet(in_features, out_features, width, depth)
     elif network_type == 'fully_connected':
@@ -58,8 +76,17 @@ def setup_network(in_features, out_features, width, depth, network_type='resnet'
     return f
 
 
-def timing_test_cpu(f, x, num_trials=10, clear_memory=True):
+def timing_test_cpu(f: Union[net.NN, torch.nn.Module], x: torch.Tensor, num_trials: int = 10,
+                    clear_memory: bool = True) -> torch.tensor():
+    r"""
+    Timing test for one architecture on CPU.
 
+    Test is run ``num_trials`` times and the timing for each trial is returned.
+
+    The timing includes one dry run for the first iteration that is not returned.
+
+    Memory is cleared after the completion of all trials.
+    """
     total_time = torch.zeros(num_trials + 1)
     for i in range(num_trials + 1):
         t1_start = time.perf_counter()
@@ -75,7 +102,19 @@ def timing_test_cpu(f, x, num_trials=10, clear_memory=True):
     return total_time[1:]
 
 
-def timing_test_gpu(f, x, num_trials=10, clear_memory=True):
[…]
