+++ hessQuik/utils/network_derivative_check.py
@@ -2,25 +2,94 @@ import torch
 from math import log2, floor
 import hessQuik
 from hessQuik.utils import convert_to_base, extract_data, insert_data
+from typing import Callable, Tuple, Optional, Union
 
 
-def network_derivative_check(f, x, do_Hessian=False, forward_mode=True, num_test=15, base=2, tol=0.1, verbose=False):
+def network_derivative_check(f: torch.nn.Module, x: torch.Tensor, do_Hessian: bool = False,
+                             forward_mode: bool = True, num_test: int = 15, base: float = 2.0, tol: float = 0.1,
+                             verbose: bool = False) -> Optional[bool]:
+    r"""
+    Taylor approximation test to verify derivatives.  Form the approximation by perturbing the
+    network weights :math:`\theta` in the direction :math:`p` with step size :math:`h > 0` via
 
+    .. math::
+
+        \Phi(\theta + h p) \approx \Phi(\theta) + h\nabla_{\theta} \Phi(\theta)^\top p
+
+    where :math:`\Phi` is the objective function and :math:`\theta` are the network weights.  This test uses the loss
+
+    .. math::
+
+        \Phi(\theta) = \frac{1}{2}\|f_{\theta}(x)\|^2 + \frac{1}{2}\|\nabla f_{\theta}(x)\|^2 +  \frac{1}{2}\|\nabla^2 f_{\theta}(x)\|^2
+
+    to validate network gradient computation after computing derivatives of the
+    input features of the network :math:`f_{\theta}`.
+
+    As :math:`h \downarrow 0^+`, the error between the approximation and the true value will decrease.
+    The rate of decrease indicates the accuracy of the derivative computation.
+    For details, see Chapter 5 of `Computational Methods for Electromagnetics`_ by Eldad Haber.
+
+    .. _Computational Methods for Electromagnetics: https://epubs.siam.org/doi/book/10.1137/1.9781611973808
+
+    Examples::
+
+        >>> import hessQuik.activations as act, hessQuik.layers as lay, hessQuik.networks as net
+        >>> torch.set_default_dtype(torch.float64)  # use double precision to check implementations
+        >>> x = torch.randn(10, 4)
+        >>> width, depth = 8, 3
+        >>> f = net.NN(lay.singleLayer(4, width, act=act.tanhActivation()), net.resnetNN(width, depth, h=1.0, act=act.tanhActivation()), lay.singleLayer(width, 1, act=act.identityActivation()))
+        >>> network_derivative_check(f, x, do_Hessian=True, verbose=True, forward_mode=True)
+            h                           E0                      E1
+            1.00 x 2^(00)		1.97 x 2^(02)		1.05 x 2^(01)
+            1.00 x 2^(-01)		1.14 x 2^(02)		1.71 x 2^(-02)
+            1.00 x 2^(-02)		1.20 x 2^(01)		1.50 x 2^(-04)
+            1.00 x 2^(-03)		1.22 x 2^(00)		1.40 x 2^(-06)
+            1.00 x 2^(-04)		1.24 x 2^(-01)		1.35 x 2^(-08)
+            1.00 x 2^(-05)		1.24 x 2^(-02)		1.32 x 2^(-10)
+            1.00 x 2^(-06)		1.24 x 2^(-03)		1.31 x 2^(-12)
+            1.00 x 2^(-07)		1.24 x 2^(-04)		1.30 x 2^(-14)
+            1.00 x 2^(-08)		1.25 x 2^(-05)		1.30 x 2^(-16)
+            1.00 x 2^(-09)		1.25 x 2^(-06)		1.30 x 2^(-18)
[…]
