### 2022-02-27 057edf1
draft of utils documentation
