### 2022-02-28 f940421
added compatible composition check
