+++ hessQuik/utils/__init__.py
@@ -1,4 +1,4 @@
-from .utils import convert_to_base, extract_data, insert_data, check_compatible_composition
+from .utils import convert_to_base, extract_data, insert_data
 from .input_derivative_check import input_derivative_check, input_derivative_check_finite_difference
 from .network_derivative_check import network_derivative_check
 from .timing import timing_test
