### 2022-02-28 a120a65
draft of networks documentation
