+++ docs/hessQuik_API/hessQuik.rst
@@ -1,21 +1,0 @@
-hessQuik package
-================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   hessQuik.activations
-   hessQuik.layers
-   hessQuik.networks
-   hessQuik.utils
-
-Module contents
----------------
-
-.. automodule:: hessQuik
-   :members:
-   :undoc-members:
-   :show-inheritance:
