+++ docs/hessQuik_setup/index.rst
@@ -1,7 +1,0 @@
-Getting Started
-===============
-
-.. toctree::
-   :maxdepth: 2
-
-   install
