+++ docs/hessQuik_functionality/index.rst
@@ -1,9 +1,8 @@
-hessQuik API
-============
+hessQuik Functionality
+======================
 
 .. toctree::
    :maxdepth: 2
-   :Caption: hessQuik API
 
    hessQuik.activations
    hessQuik.layers
