+++ docs/hessQuik_setup/install.rst
@@ -4,3 +4,11 @@ Installation
 Use ``pip`` from the command line::
 
     python -m pip install git+https://github.com/elizabethnewman/hessQuik.git
+
+
+
+The following are the ``hessQuik`` dependencies:
+
+    - ``torch`` (with the recommended version >= 1.10.0, but code will run with version >= 1.5.0)
+
+The dependencies are installed automatically with ``pip``.
