+++ .github/workflows/python-package.yml
@@ -31,4 +31,4 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Test with pytest
       run: |
-        pytest ../tests/test_activations.py
+        pytest test_activations.py
