+++ .github/workflows/python-app.yml
@@ -24,7 +24,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        pip install torch -f https://download.pytorch.org/whl/torch_stable.html
+        pip3 install torch
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Lint with flake8
       run: |
