+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-testpaths = hessQuik/tests
+testpaths = tests
 python_files = test_*.py
 addopts = -rf --import-mode=importlib
