+++ .github/workflows/python-app.yml
@@ -26,7 +26,7 @@ jobs:
         pip install flake8 pytest
         pip3 install torch
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        python -m pip install git+https://github.com/elizabethnewman/hessQuik.git
+        pip install .
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
