+++ .github/workflows/python-app.yml
@@ -26,7 +26,7 @@ jobs:
         pip install flake8 pytest
         pip3 install torch
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-        pip install .
+        pip install -e .
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
