+++ hessQuik/layers/single_layer.py
@@ -150,8 +150,9 @@ class singleLayer(hessQuikLayer):
 
                 # extra term to compute full Hessian
                 h2 = d2gd2u @ dgdf.unsqueeze(1)
+
                 # combine
-                d2gd2u = h1 + 0 * h2
+                d2gd2u = h1 + h2
 
         # finish computing gradient
         if dgdf is not None:
