+++ hessQuik/layers/single_layer.py
@@ -113,7 +113,7 @@ class singleLayer(hessQuikLayer):
         if (do_gradient or do_Hessian) and forward_mode is False:
             dfdx, d2fd2x = self.backward(do_Hessian=do_Hessian)
 
-        return f, dfdx, d2fd2x
+        return f, 0 * dfdx, d2fd2x
 
     def backward(self, do_Hessian: bool = False,
                  dgdf: Union[torch.Tensor, None] = None, d2gd2f: Union[torch.Tensor, None] = None):
