### 2022-02-28 0942814
changed data type for tests
