+++ hessQuik/tests/test_activations.py
@@ -8,37 +8,44 @@ class TestActivation(unittest.TestCase):
 
     @staticmethod
     def run_test(f):
+        torch.set_default_dtype(torch.float64)
         nex = 11
         d = 7
         x = torch.randn(nex, d)
         run_all_tests(f, x)
 
     def test_antiTanhActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.antiTanhActivation()
         print(type(f))
         self.run_test(f)
 
     def test_quadraticActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.quadraticActivation()
         print(type(f))
         self.run_test(f)
 
     def test_softplusActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.softplusActivation()
         print(type(f))
         self.run_test(f)
 
     def test_identityActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.identityActivation()
         print(type(f))
         self.run_test(f)
 
     def test_tanhActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.tanhActivation()
         print(type(f))
         self.run_test(f)
 
     def test_sigmoidActivation(self):
+        torch.set_default_dtype(torch.float64)
         f = act.sigmoidActivation()
         print(type(f))
         self.run_test(f)
