### 2022-02-28 elizabethnewman/hessQuik#2 @elizabethnewman
Dear @yhtang and @GregaVrbancic

Thank you for your suggestions and for reviewing this repository.  As requested, I have added new API functionality documentation, linked in the README, and have added a continuous integration workflow using Github Actions.  I believe these changes make the repository much more user-friendly and fulfill the JOSS requirements (I hope!).  Please let me know if there is more I can do to improve the documentation.
