### 2022-02-28 elizabethnewman/hessQuik#2 @yhtang
Thanks for the update! Given that the overall structure of the documentation is largely in place now, I'm going to close this issue. We can then fall back to the main review thread and address any remaining issues there.
