### 2022-03-03 elizabethnewman/hessQuik#4 @GregaVrbancic
[JOSS review] Packaging
Hi @elizabethnewman,

firstly, I would like to thank you for a great job with documentation. Here is just a small suggestion regarding the packaging of the library. I would suggest you upload the package to pypi. This would enable users to find and install the library more easily. Again, this is in my opinion not mandatory, but it would be nice to have 🙂
