### 2022-03-03 elizabethnewman/hessQuik#5 @GregaVrbancic
[JOSS review] Community guidelines
Hi @elizabethnewman, could you please add some clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

If the guidelines are already written somewhere, I would be glad if you could provide me with the link to the guidelines.
