### 2022-03-03 #4171 @yhtang
@elizabethnewman: I'm having some difficulties running the [timing result notebook](https://github.com/elizabethnewman/hessQuik/blob/main/hessQuik/examples/hessQuikTimingTest.ipynb) on my local machine. Specifically, the problem seems to be related to invoking system commands (e.g. `!python`) in a notebook cell. In that case, the system's default Python interpreter gets called, instead of the one in the virtual environment.
