### 2022-03-07 aec351e
added contributing documentation
