+++ docs/hessQuik_setup/contributing.rst
@@ -1,0 +1,34 @@
+Contributing
+============
+
+There are many ways to contribute to ``hessQuik``.
+
+Contributing to the software
+----------------------------
+
+1. Fork the ``hessQuik`` repository
+2. Clone your fork using the command::
+
+    git clone https://github.com/<username>/hessQuik.git
+
+3. Contribute to your forked repository.
+4. Create a pull request.
+
+If your code passes the necessary numerical tests and is well-documented, your changes and/or additions will be merged in the main ``hessQuik`` repository.
+
+You can find examples of the tests used in each file and related unit tests the tests directory.
+
+
+Reporting Issues
+----------------
+
+If you notice an issue with this repository, please report it using `Github Issues`_. When reporting an implementation bug, include a small example that helps to reproduce the error. The issue will be addressed as quickly as possible
+
+.. _Github Issues: https://docs.github.com/en/issues
+
+Seeking Support
+---------------
+
+If you have questions or need additional support, please open a `Github Issue`_ or send a direct email to elizabeth.newman@emory.edu.
+
+.. _Github Issue: https://docs.github.com/en/issues
