### 2022-03-07 elizabethnewman/hessQuik#5 @elizabethnewman
Hi @GregaVrbancic

Thank you for your question.  I had included community guidelines in the repository README file, in the Contributing and Reporting Bugs sections.  To make this easier to find, I have added the contributing directions to the documentation as well: https://hessquik.readthedocs.io/en/latest/hessQuik_setup/contributing.html.  I hope this addresses your question!
