### 2022-03-07 #4171 @elizabethnewman
>

Hi @yhtang

Thank you for running the code.  To ensure reproducibility, timing result notebook is designed to be run on Google Colab only.  However, the script run_timing_test.py, which is called in the notebook, should be compatible with your local machine.  I have clarified the instructions in the examples directory README file.  I hope this helps!
