+++ docs/hessQuik_setup/install.rst
@@ -1,11 +1,13 @@
 Installation
 ============
 
-Use ``pip`` from the command line::
+To install from PyPI, use the following from the command line::
+
+    pip install hessQuik
+
+To install from Github, use the following from the command line::
 
     python -m pip install git+https://github.com/elizabethnewman/hessQuik.git
-
-
 
 The following are the ``hessQuik`` dependencies:
 
