### 2022-03-07 elizabethnewman/hessQuik#4 @elizabethnewman
Hi @GregaVrbancic

Thank you for your suggestion about uploading the package to PyPI.  I have done so and added instructions to the Github README and to the documentation about importing from PyPI.  I hope this makes installation easier!
