### 2022-03-08 #4171 @yhtang
> However, the script run_timing_test.py, which is called in the notebook, should be compatible with your local machine. I have clarified the instructions in the examples directory README file. I hope this helps!

I'll get it a try. Thanks!

> Thank you for running the code. To ensure reproducibility, timing result notebook is designed to be run on Google Colab only.

This actually brings up another issue that I wanted to ask while reading through the paper. Since Colab assigns GPUs based on resource availability (and in my own experience we can potentially get anything from a decade-old K80 to a newer model like V100), I'm not quite sure that it is the best platform to showcase reproducibility --- at least not so quantitatively. Someone would have to be a bit lucky in order to obtain the same hardware resource. This was also part of my motivation to run the notebook locally.

Perhaps it'll be a good idea to clarify and mention these subtleties in the paper?
