### 2022-03-08 1412e6c
new local timing test notebook
