+++ hessQuik/examples/README.md
@@ -29,5 +29,7 @@ python run_timing_test.py --num-input 5 --num-output 4 --network-wrapper hessQui
 ```
 The above command runs the timing test with our hessQuik implementation for 5 different input feature sizes from $2^0$ to $2^4$ and 4 different output features sizes from $2^0$ to $2^3$.
 
-The flags used in the paper experiments are available in the notebook
+The flags used in the paper experiments are available in the notebook:
+
+[HessQuikTimingTestLocal.ipynb](https://github.com/elizabethnewman/hessQuik/blob/main/hessQuik/examples/hessQuikTimingTestLocal.ipynb)
 
