### 2022-03-28 #4171 @GregaVrbancic
Dear @jbytecode and @elizabethnewman, I have finished my review. In addition to already adequately addressed issues https://github.com/elizabethnewman/hessQuik/issues/4 and https://github.com/elizabethnewman/hessQuik/issues/5 I have no further concerns regarding the package or paper. Therefore, I would recommend this submission for publication.
