### 2022-04-22 8ed30ed
corrections due to the JOSS submission
