### 2022-04-22 #4171 @jbytecode
Dear @elizabethnewman

I have just created a pull request (https://github.com/elizabethnewman/hessQuik/pull/6) in the software repository.

Please review them carefully and apply them if you are agreed with the corrections.

Thank you in advance.
