### 2022-04-22 #4171 @elizabethnewman
> Dear @elizabethnewman
>
> It seems okay now.
>
> * [x]  Please read the paper pdf carefully, one more time.
> * [x]  Create a tagged release in the software repository. Report the version here. The version should be in form of `vx.y.z`, e.g. `v1.2.3`.
> * [x]  Create a Zenodo archive in https://zenodo.org/
> * [x]  The Zenodo archive should have the correct metadata, e.g. author names, ORCIDs, the title of the archive. Despite it is not a rigid requirement, please set the archive title as the paper title.
> * [x]  Please provide the tagged release version id, zenodo archive DOI, zenodo archive url here.
>
> Thank you in advance.

Dear @jbytecode

I believe I have followed all of the steps you have outlined above.  Here is the requested information:

tagged release version id: v1.0.0
zenodo archive DOI: 10.5281/zenodo.6478757
zenodo archive url: https://zenodo.org/record/6478757#.YmMyAfPMK3J

Please let me know if you need any other information.  Thank you!
