### 2022-04-23 #4171 @arfon
I *think* the issue is that this is actually an `ISSN` not `ISBN`: https://github.com/elizabethnewman/hessQuik/blob/main/paper/paper.bib#L179 and so the current value is failing the XML validation.
