### 2022-04-24 #4171 @jbytecode
Dear @elizabethnewman

I have just send a new pull request (https://github.com/elizabethnewman/hessQuik/pull/7) that possibly fixes the compilation task. Could you please review and apply it?

Thank you.
